Add burger menu component
- Add package react-burger-menu
- Add new headerNav component
- Remove a menu component
- Change the estructure of the header
