+++ package.json
@@ -33,6 +33,7 @@
     "graphql": "^14.3.1",
     "node-sass": "^4.12.0",
     "react": "^16.8.6",
+    "react-burger-menu": "^2.6.10",
     "react-dom": "^16.8.6",
     "react-helmet": "^5.2.1",
     "react-icons": "^3.7.0"
