+++ src/components/header.module.scss
@@ -14,13 +14,7 @@
   color: white;
   margin-left: 2vw;
 }
-.navigation {
-  display: flex;
-  align-items: center;
-  height: 100%;
-  // padding: 20px 0px;
-  justify-content: space-between;
-}
+
 .sitenames {
   font-size: 1.5rem;
   margin:0;
