+++ src/components/headerNav.module.scss
@@ -1,0 +1,52 @@
+@import '../styles/style.scss';
+
+.navigation {
+  display: flex;
+  align-items: center;
+  height: 100%;
+  // padding: 20px 0px;
+  justify-content: space-between;
+}
+
+.active {
+  display: inline;
+}
+
+/* Add a black background color to the top navigation */
+.topnav {
+  background-color: white;
+  overflow: hidden;
+  text-align: right;
+  flex-grow: 8;
+  /* Style the links inside the navigation bar */
+  a {    
+    display: inline-block;
+    // color: $nav-items-color;
+    text-align: center;
+    // padding: 14px 16px;
+    padding: 0 1vw;
+    text-decoration: none;
+    font-size: 17px;
+  }
+  
+  /* Change the color of links on hover */
+  a:hover {    
+    color: $primary-color;
+    // color: $nav-items-color-hover;
+  }
+  /* Add an active class to highlight the current page */
+  a.active {    
+    // color:$nav-items-color-active;
+    font-weight: bolder;
+  }
+  a.bars {
+    display: none;
+  }
+}
+
+@media screen and (max-width: 768px) {
+  // .topnav a:not(:first-child) {display: none;}
+  .topnav a {
+    display: none;
+  }
+}
