+++ src/components/layout.js
@@ -1,20 +1,15 @@
 import React from 'react'
-import Header from './header'
 import Footer from './footer'
 import PropTypes from 'prop-types'
-// import '../styles/index.scss'
 import '../styles/style.scss'
-// import layoutStyles  from './layout.module.scss'
+import BurgerMenu from '../components/burgerMenu'
 
 const Layout = (props) => {
   return (
-    // <div className={layoutStyles.container}>
     <div className="">
-      {/* <div className={layoutStyles.content}> */}
-      <div className="">
-        <Header />
-        {props.children}
-      </div>
+      <BurgerMenu />
+      {props.children}
+      
       <Footer/>
     </div>
   )
