--- src/components/menu.module.scss
@@ -1,75 +1,0 @@
-// @import '../styles/settings.scss';
-@import '../styles/style.scss';
-
-.title {
-  color: $primary-color;
-  text-decoration: none;
-  display: flex;
-  float:left;
-}
-
-/* Add a black background color to the top navigation */
-.topnav {
-  background-color: white;
-  overflow: hidden;
-  text-align: right;
-  flex-grow: 8;
-  /* Style the links inside the navigation bar */
-  a {    
-    display: inline-block;
-    // color: $nav-items-color;
-    text-align: center;
-    // padding: 14px 16px;
-    padding: 0 1vw;
-    text-decoration: none;
-    font-size: 17px;
-  }
-  
-  /* Change the color of links on hover */
-  a:hover {    
-    color: $primary-color;
-    // color: $nav-items-color-hover;
-  }
-  /* Add an active class to highlight the current page */
-  a.active {    
-    // color:$nav-items-color-active;
-    font-weight: bolder;
-  }
-  a.bars {
-    display: none;
-  }
-}
-
-/* When the screen is less than 768 pixels wide, hide all links, except for the first one ("Home"). 
-Show the link that contains should open and close the topnav (.icon) */
-@media screen and (max-width: 768px) {
-  // .topnav a:not(:first-child) {display: none;}
-  .topnav a {
-    display: none;
-  }
-  .topnav a.bars {
-    float: right;
-    display: block;
-    display: flex;
-    align-items: center;
-    height: 120px;
-  }
-}
-
-/* The "responsive" class is added to the topnav with JavaScript when the user clicks on the icon. 
-This class makes the topnav look good on small screens (display the links vertically instead of horizontally) */
-@media screen and (max-width: 768px) {
-  .topnav.responsive {position: relative;}
-  .topnav.responsive a.bars {
-    position: absolute;
-    right: 0;
-    top: 0;
-    display: flex;
-    align-items: center;
-  }
-  .topnav.responsive a {
-    float: none;
-    display: block;
-    text-align: left;
-  }
-}
