+++ src/styles/style.scss
@@ -4,6 +4,7 @@
 @import '~bulma/bulma.sass';
 @import '~bulma-divider/dist/css/bulma-divider.sass';
 
+@import './burgerMenu.scss';
 
 .site-name{
   color: $logo-letter-color !important;
